python/pyproject.toml
[project]
dependencies = [
  "lark-parser==0.12.0"
]
name = "UNKNOWN"
version = "0.0.0"

[project.optional-dependencies]
dev = [
  "coverage==7.3.2",
  "djlint==1.34.0",
  "mypy==1.6.1",
  "ruff==0.1.3",
  "ssort==0.11.6",
  "toml-sort==0.23.1"
]
